requirements.txt
pandas>=1.5
numpy>=1.23
faiss-cpu>=1.7
openai>=1.0.0
flask>=2.0
flask-cors>=3.0